messy-migration/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Flask-SQLAlchemy>=3.0
marshmallow>=3.0
marshmallow-sqlalchemy>=0.27
python-dotenv>=0.21
pytest>=7.0
